\name{agres}
\alias{agres}
\docType{data}
\title{European agricultural data}
\description{Data on research expenditure, research activity, productivity and impacts of Agriculture from 1990 to 2014 in EU 15 countries.}
\usage{data(agres)}
\format{
  A data frame with a total of 350 observations on the following 20 variables:
  \describe{
    \item{\code{COUNTRY}}{Code of the country. Belgium and Luxembourg are merged together with code \code{BL}.}
    \item{\code{YEAR}}{Year.}
    \item{\code{GDP}}{Gross domestic product (million euro PPS). Source: Eurostat.}
    \item{\code{EMPL_AGR}}{Persons employed in Agriculture (count). Source: Eurostat.}
    \item{\code{UAA}}{Utilized agricultural area (hectares). Source: Eurostat.}
    \item{\code{PATENT_OTHER}}{Mechanical, chemical and environment-related patent applications (count). Source: OECD.}
    \item{\code{GBAORD_AGR}}{Government research expenditure (million euro PPS). Source: OECD.}
    \item{\code{BERD_AGR}}{Business enterprise research expenditure (million euro PPS). Source: Eurostat.}
    \item{\code{RD_EDU_AGR}}{Agricultural researchers with tertiary education (count). Source: OECD.}
    \item{\code{PATENT_AGR}}{Agricultural patent applications (count). Source: OECD.}
    \item{\code{TFP_AGR}}{Total factor productivity (index 2005=100). Source: USDA.}
    \item{\code{GVA_AGR}}{Gross value added of Agriculture (international dollars). Source: Faostat.}
    \item{\code{PPI_AGR}}{Producer price of agricultural output (index 2005=100). Source: Faostat.}
    \item{\code{ENTR_INCOME_AGR}}{Net entrepreneurial income of Agriculture (index 2005=100). Source: Eurostat.}
    \item{\code{GHG_AGR}}{Greenhouse gas emissions due to Agriculture (gigagrams CO2 equivalent). Source: Faostat.}
    \item{\code{NBAL_AGR}}{Nitrogen gross nutrient balance (kg of nutrient per hectare). Source: Eurostat.}
    \item{\code{INCOME_RURAL}}{Mean familiar income in rural areas (international dollars).}
    \item{\code{UNEMPL_RURAL}}{Unemployment rate in rural areas (index 2005=100).}
  }
}
