\name{industry}
\alias{industry}
\docType{data}
\title{Industrial development}
\description{Fictitious data on industrial development from 1983 to 2014 in 10 imaginary regions.}
\usage{data(industry)}
\format{
  A data frame with a total of 320 observations on the following 7 variables:
  \describe{
    \item{\code{Region}}{ID of the region.}
    \item{\code{Year}}{Year.}
    \item{\code{Population}}{Population (number of inhabitants).}
    \item{\code{GDP}}{Gross domestic product (1000 international dollars).}
    \item{\code{Job}}{Employment in industry (Index, 1980=100).}
    \item{\code{Consum}}{Private consumption (Index, 1980=100).}
    \item{\code{Pollution}}{Greenhouse gas emissions (tonnes of CO2 equivalent).}
  }
}
