\name{agres}
\alias{agres}
\docType{data}
\title{Agricultural research data}
\description{Data on agricultural research, profitability and consumer surplus for 10 European countries in the period 1990-2012.}
\usage{data("agres")}
\format{
  A data frame with a total of 260 observations on the following 8 variables:
  \describe{
    \item{\code{COUNTRY}}{Country (international code).}
    \item{\code{YEAR}}{Year.}
    \item{\code{GDP}}{Gross domestic product.}
    \item{\code{FARM_SIZE}}{Average farm size: number of persons employed in Agriculture to total hectares of utilized agricultural area.}
    \item{\code{NPATENT}}{Number of agricultural patent applications.}
    \item{\code{GVA}}{Gross value added of Agriculture.}
    \item{\code{PPI}}{Producer price of agricultural output (Index, 1980=100).}
    \item{\code{ENTR_INCOME}}{Net intrepreneural income of Agriculture (Index, 2005=100).}
  }
}
\source{EUROSTAT database. \url{http://ec.europa.eu/eurostat/data/database}}
