\name{dlsem-package}
\alias{dlsem-package}
\docType{package}
\title{Distributed-lag structural equation modelling}
\description{Inference functionalities for structural equation models with constrained lag shapes (Magrini et al., 2016).
The main functions of the package are:
\itemize{
  \item{\link{dlsem}, }{to perform parameter estimation;}
  \item{\link{causalEff}, }{to assess and disentangle causal effects at different time lags;}
  \item{\link{lagPlot}, }{to plot the lag shape associated to a path or to an overall causal effect.} 
  }
}
\details{
  \tabular{ll}{
  Package: \tab dlsem\cr
  Type: \tab Package\cr
  Version: \tab 1.7\cr
  Date: \tab 2017-01-10\cr
  License: \tab GPL-2
  }
}
%\author{Alessandro Magrini <magrini@disia.unifi.it>}
\references{A. Magrini, F. Bartolini, A. Coli, and B. Pacini (2016). Distributed-Lag Structural Equation Modelling:
An Application to Impact Assessment of Research Activity on European Agriculture.
\emph{Proceedings of the 48th Meeting of the Italian Statistical Society}, 8-10 June 2016, Salerno, IT.}        
