\name{dlsem-package}
\alias{dlsem-package}
\docType{package}
\title{Distributed-lag structural equation modelling with second-order polynomial lag shapes}
\description{Estimation and path analysis functionalities for structural equation models with second-order polynomial lag shapes.}
\details{
  \tabular{ll}{
  Package: \tab dlsem\cr
  Type: \tab Package\cr
  Version: \tab 1.3.2\cr
  Date: \tab 2016-09-16\cr
  License: \tab GPL-2
  }
}
\author{Alessandro Magrini <magrini@disia.unifi.it>}
\references{A. Magrini, F. Bartolini, A. Coli, and B. Pacini (2016). Distributed-Lag Structural Equation Modelling:
An Application to Impact Assessment of Research Activity on European Agriculture.
\emph{Proceedings of the 48th Meeting of the Italian Statistical Society}, 8-10 June 2016, Salerno, IT.}        
